Enum menhir_runtime::Error

pub enum Error<Lexer: Lexer, Parser: LRParser> {
    SyntaxError(SyntaxError<Lexer, Parser>),
    LexerError(Lexer::Error),
}

A fatal (non-recoverable parsing error).

This is the type returned by a Menhir parser upon encountering an error it cannot or failed to recover from.

Variants
SyntaxError(SyntaxError<Lexer, Parser>)

The parser encountered a syntax error that couldn't be recovered. See SyntaxError.

LexerError(Lexer::Error)

The lexer returned an error when asked for a token, typically an IO error or the end of the stream. The carried value depends on the actual instance of Lexer being used.