Struct menhir_runtime::lexing::IteratorLexer

pub struct IteratorLexer<Iter, Loc, Tok> where
    Iter: Iterator<Item = (Loc, Tok)>,  { /* fields omitted */ }

Adapter type to convert an iterator into a lexer.

Lexing tools that implement an interface based on Iterator can be converted to Menhir's more advanced Lexer interface by using this type.

If the iterator returns None while the parser still needs tokens, input will return an UnexpectedEof error. If one wants the parser to consume exactly all the tokens, one should manually check that the underlying iterator is empty once parsing succeeded.

Methods

impl<Iter, Loc, Tok> IteratorLexer<Iter, Loc, Tok> where
Loc: Default,
Iter: Iterator<Item = (Loc, Tok)>,

fn new(lex: Iter) -> Self

Builds a new adapter from the given iterator.

This function takes the iterator by-value. If the iterator is not, Copy, it won't be usable anymore after the parsing process. If this is not desired, one should give to this function a mutable reference to the iterator instead.