Struct menhir_runtime::ParserState

pub struct ParserState<Lexer, Parser, Entry> where
    Parser: LRParser,
    Lexer: Lexer,
    Entry: EntryPoint<Parser>,  { /* fields omitted */ }

The state of the parser during the parsing process.

The interface of this type should not be considered stable. In the future, this type will be used to expose the incremental interface, which allows to run the parsing process step-by-step and access its internals

Methods

impl<Lexer, Parser, Entry> ParserState<Lexer, Parser, Entry> where
Parser: LRParser,
Lexer: Lexer,
Lexer::Token: Into<(Parser::YYType, Parser::Terminal)>,
Lexer::Location: Clone,
Entry: EntryPoint<Parser>,

fn run(self, lexer: Lexer) -> Result<Entry::Output, Error<Lexer, Parser>>

Run the parsing process until success or failure.

This function actually implements the EntryPoint::run function in terms of the internal, unstable incremental interface. See EntryPoint::run for more details about the behaviour of this function.