Struct menhir_runtime::SyntaxError

pub struct SyntaxError<Lexer: Lexer, Parser: LRParser> { /* fields omitted */ }

An unrecoverable syntax error.

Allows to extract the position in the input stream and the automaton state in which the error was detected.

Methods

impl<Parser: LRErrors, Lexer: Lexer> SyntaxError<Lexer, Parser>

fn as_str(&self) -> Option<&'static str>

Returns a propre error message to describe this error, if such a message was specified.

This require the parser to implement LRErrors, which is the case when the grammar is compiled with the --compile-error option. See the manual for information about how to use this feature.

impl<Parser: LRParser, Lexer: Lexer> SyntaxError<Lexer, Parser>

fn location(&self) -> &Lexer::Location

Returns the location in the input stream of the token where the error was detected.

The actual type of this value depends on the instance of Lexer being used.